==> notebooks/01_merge_datasets.ipynb <==
import pandas as pd
import glob


# ENROLMENT
files = glob.glob("../data/raw/enrolment/*.csv")
df_enrol = pd.concat([pd.read_csv(f) for f in files], ignore_index=True)
df_enrol.to_csv("../data/processed/enrolment_full.csv", index=False)


# DEMOGRAPHIC
files = glob.glob("../data/raw/demographic/*.csv")
df_demo = pd.concat([pd.read_csv(f) for f in files], ignore_index=True)
df_demo.to_csv("../data/processed/demographic_full.csv", index=False)



# BIOMETRIC
files = glob.glob("../data/raw/biometric/*.csv")
df_bio = pd.concat([pd.read_csv(f) for f in files], ignore_index=True)
df_bio.to_csv("../data/processed/biometric_full.csv", index=False)

print("All datasets merged successfully!")

print("ENROLMENT COLUMNS:\n", df_enrol.columns)
print("\nDEMOGRAPHIC COLUMNS:\n", df_demo.columns)
print("\nBIOMETRIC COLUMNS:\n", df_bio.columns)
